chore: complete the image renaming
